+++ src/lib/edge-routing-scheme17.test.ts
@@ -109,7 +109,6 @@ function makeNode(
       machineCount: 1,
       overclock: 1,
       parallel: 1,
-      outputMultiplier: 1,
       onRecipeChange: () => {},
       onMachineCountChange: () => {},
       onOverclockChange: () => {},