+++ src/lib/edge-routing-scheme18.test.ts
@@ -107,7 +107,6 @@ function makeNode(
       machineCount: 1,
       overclock: 1,
       parallel: 1,
-      outputMultiplier: 1,
       onRecipeChange: () => {},
       onMachineCountChange: () => {},
       onOverclockChange: () => {},