+++ src/lib/edge-routing-scheme46.test.ts
@@ -87,7 +87,6 @@ function makeNode(
       machineCount: 1,
       overclock: 1,
       parallel: 1,
-      outputMultiplier: 1,
       onRecipeChange: () => {},
       onMachineCountChange: () => {},
       onOverclockChange: () => {},